Tests/AK: Merge pointer tests
This commit merges pointer tests together. They fit together well, and
there is no reason for them to be split.
